Install

+++ install.sh
@@ -14,7 +14,7 @@ git stash drop
 
 # mongo
 apt-get update
-if [ ! -f "/etc/apt/sources.list.d/mongodb-org-4.4.list"]; then
+if [ ! -f "/etc/apt/sources.list.d/mongodb-org-4.4.list" ]; then
   apt-get install -y gnupg wget
   wget -qO - https://www.mongodb.org/static/pgp/server-4.4.asc | apt-key add -
   UBUNTU_VERISON=$(cat /etc/os-release | awk -F '=' '/^VERSION_ID/{print $2}' | awk '{print $1}' | tr -d '"')
@@ -33,17 +33,17 @@ if [ ! -f "/etc/apt/sources.list.d/mongodb-org-4.4.list"]; then
   apt-get update
 fi
 apt-get install -y mongodb-org
-
 systemctl enable mongod
 systemctl start mongod
 
-
 # node
-apt-get install -y curl build-essential
-curl -sL https://deb.nodesource.com/setup_12.x -o nodesource_setup.sh
-chmod +x nodesource_setup.sh
-./nodesource_setup.sh
-apt-get install -y nodejs
+if [ ! -f nodesource_setup.sh ]; then
+  apt-get install -y curl
+  curl -sL https://deb.nodesource.com/setup_12.x -o nodesource_setup.sh
+  chmod +x nodesource_setup.sh
+  ./nodesource_setup.sh
+fi
+apt-get install -y nodejs build-essential
 npm install -g npm@latest
 pushd back
 npm install
@@ -57,6 +57,6 @@ popd
 # inno-comp
 systemctl stop inno-comp.service
 systemctl disable inno-comp.service
-sed -i "s/#DIRECTORY#/$PWD/" inno-comp.service
+sed -i "s/$$DIRECTORY/$PWD/" inno-comp.service
 systemctl enable inno-comp.service
 systemctl start inno-comp.service
